Adjusting small application points

+++ src/Service/Domain/Interfaces/IProcess.cs
@@ -1,9 +1,9 @@
 ﻿namespace Service.Domain.Interfaces;
 public interface IProcess
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public Guid CycleId { get; set; }
     public bool IsExecuted { get; set; }
-    
+
     public void SetExecuted();
 }
